refactor(step2): 클래스 이름 및 디렉토리 경로 변경
